refactor(memory): rename internal modules to public names matching importer reality
The five mindroom.memory._X internal modules (_file_backend, _mem0_backend,
_policy, _prompting, _shared) all have production importers under src/
(intra-package siblings, plus mindroom.memory facade for _prompting). Per the
naming convention "private iff nothing under src/ imports it", they are
de-facto public and should drop the leading underscore. Tach module-level
visibility=[…] whitelists already enforce the real access boundary.

Also adds the missing visibility whitelist to mindroom.memory.policy.
