+++ src/mindroom/memory/shared.py
